requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
mutagen==1.47.0
py-cord==2.6.1
pycryptodomex==3.20.0
python-mpv==1.0.7
requests==2.32.3
urllib3==2.2.3
websockets==13.1
yarl==1.13.1
yt-dlp==2024.9.27
